third_party/blink/web_tests/external/wpt/mathml/presentation-markup/mrow/no-spacing-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Spacing around an <mfrac> child of an <mtd>.
  assert_greater_than_equal: expected a number greater than or equal to 200 but got 47
[FAIL] Spacing around an <msub> child of an <mtd>.
  assert_greater_than_equal: expected a number greater than or equal to 200 but got 75
[FAIL] Spacing around an <msup> child of an <mtd>.
  assert_greater_than_equal: expected a number greater than or equal to 200 but got 75
[FAIL] Spacing around an <msubsup> child of an <mtd>.
  assert_greater_than_equal: expected a number greater than or equal to 200 but got 75
[FAIL] Spacing around an <mmultiscripts> child of an <mtd>.
  assert_greater_than_equal: expected a number greater than or equal to 200 but got 105
[FAIL] Spacing around an <munder> child of an <mtd>.
  assert_greater_than_equal: expected a number greater than or equal to 200 but got 45
[FAIL] Spacing around an <mover> child of an <mtd>.
  assert_greater_than_equal: expected a number greater than or equal to 200 but got 45
[FAIL] Spacing around an <munderover> child of an <mtd>.
  assert_greater_than_equal: expected a number greater than or equal to 200 but got 45
Harness: the test ran to completion.
